requirements.txt
openai==1.101.0
pymupdf==1.26.4
PyPDF2==3.0.1
python-dotenv==1.1.0
tiktoken==0.11.0
pyyaml==6.0.2
fastapi==0.115.12
uvicorn==0.34.3
streamlit==1.54.0
pytest==8.4.1
pytest-asyncio==1.0.0
